adding more art ... making a 4th sheet

+++ series-b/art/src/get.sh
@@ -1,5 +1,9 @@
 #!/bin/bash
 wget -nc https://upload.wikimedia.org/wikipedia/commons/6/69/Utagawa_Toyoharu_-_A_Winter_Party_-_Google_Art_Project.jpg
 wget -nc https://upload.wikimedia.org/wikipedia/en/5/54/David_Alfaro_Siqueiros_-_Peasants_-_Google_Art_Project.jpg
-wget -O christinas_world.jpg -nc "http://wallpaperbeta.com/wallpaper_3840x2160/andrew_wyeth_christina's_world_picture_stuff_ultra_3840x2160_hd-wallpaper-328903.jpg"
+[ -e christinas_world.jpg ] || wget -O christinas_world.jpg -nc "http://wallpaperbeta.com/wallpaper_3840x2160/andrew_wyeth_christina's_world_picture_stuff_ultra_3840x2160_hd-wallpaper-328903.jpg"
 wget -nc https://upload.wikimedia.org/wikipedia/commons/7/7e/Henri_Rousseau_010.jpg
+[ -e barrau_solitude.jpg ] || wget -O barrau_solitude.jpg -nc https://upload.wikimedia.org/wikipedia/en/2/22/Laure%C3%A0_Barrau_-_Solitude_-_Google_Art_Project.jpg
+[ -e the_concert ] || wget -O the_concert.jpg -nc https://upload.wikimedia.org/wikipedia/en/f/f5/Albert_Andr%C3%A9_-_The_Concert_-_Google_Art_Project.jpg
+wget -nc https://upload.wikimedia.org/wikipedia/en/0/0e/Gaspar_Homar_-_Picnic_in_the_country_-_Google_Art_Project.jpg
+[ -e cariani-a-concert.jpg ] || wget -O cariani-a-concert.jpg -nc https://upload.wikimedia.org/wikipedia/commons/8/82/Cariani_-_A_Concert_-_Google_Art_Project.jpg
